backend/requirements.txt
gradio
PyMuPDF
requests
python-dotenv
fastapi
uvicorn
pyjwt